+++ keystore/store.go
@@ -14,7 +14,7 @@ func init() {
 }
 
 func Getkey(key string) string {
-	log.Println("get key: ", key)
+	log.Println("get key:", key)
 	value, ok := keyStoreMap[key]
 	if ok {
 		return value
@@ -26,11 +26,11 @@ func Getkey(key string) string {
 
 func Addkey(Keyname string, Value string) int {
 	if _, isKeyExists := keyStoreMap[Keyname]; isKeyExists {
-		log.Println("key already exists: " + Keyname)
+		log.Println("key already exists:" + Keyname)
 		return 409
 	} else {
 		keyStoreMap[Keyname] = Value
-		log.Println("add key: ", Keyname)
+		log.Println("add key:", Keyname)
 		return 200
 	}
 }
